+++ packages/gamut/src/Form/types.ts
@@ -1,0 +1,4 @@
+export type FormValues = Record<
+  string,
+  boolean | string | undefined | FileList
+>;
